Define <Plug>(vimrc#key*) keymapping mechanism
This mechanism makes it possible to override supported keys in various
plugins while falling back to "default" behavior that can be customized
by other plugins. For example, a snippet plugin could intercept <Tab>
and expand a snippet, falling back to inserting a <Tab>, and another
plugin could redefine what it means to insert a <Tab>.

Use the new delimitMate User autocommands to integrate delimitMate with
this new keymapping behavior.